refactor: move directory_search implementation to __init__.py
