+++ agentstack/tools/directory_search/config.json
@@ -4,7 +4,7 @@
   "tools": ["search_directory", "search_fixed_directory"],
   "description": "Search through files in a directory using embedchain's DirectoryLoader",
   "env": {
-    "DIRECTORY_SEARCH_PATH": "Optional. Path to directory for fixed directory search."
+    "DIRECTORY_SEARCH_TOOL_PATH": "Optional. Path to directory for fixed directory search."
   },
   "dependencies": [
     "embedchain>=0.1.0"
